+++ src/canister/src/logic/friend_request_logic.rs
@@ -19,7 +19,7 @@ impl FriendRequestCalls {
         to: Principal,
         message: String,
     ) -> Result<FriendRequestResponse, ApiError> {
-        let mut friend_request = FriendRequest::new(caller(), to, message);
+        let friend_request = FriendRequest::new(caller(), to, message);
         let (_, caller_profile) = ProfileStore::get(caller())?;
 
         if caller_profile.relations.contains_key(&to) {
@@ -51,15 +51,25 @@ impl FriendRequestCalls {
         }
 
         // make a notification for the friend request
-        let notification_id =
-            NotificationCalls::notification_add_friend_request(friend_request.clone())?;
 
         // add the notification id to the friend request as reference
-        friend_request.set_notification_id(notification_id);
 
         // insert the friend request into the store
-        let friend_request_result = FriendRequestStore::insert(friend_request);
-        FriendRequestMapper::to_result_response(friend_request_result)
+        let (friend_request_id, mut inserted_friend_request) =
+            FriendRequestStore::insert(friend_request)?;
+
+        let friend_request_response =
+            FriendRequestResponse::new(friend_request_id, inserted_friend_request.clone());
+
+        let notification_id =
+            NotificationCalls::notification_add_friend_request(friend_request_response)?;
+
+        inserted_friend_request.set_notification_id(notification_id);
+
+        FriendRequestMapper::to_result_response(FriendRequestStore::update(
+            friend_request_id,
+            inserted_friend_request,
+        ))
     }
 
     pub fn accept_friend_request(friend_request_id: u64) -> Result<bool, ApiError> {
@@ -78,15 +88,20 @@ impl FriendRequestCalls {
             RelationType::Friend.to_string(),
         );
 
-        let (_, mut to_profile) = ProfileStore::get(friend_request.requested_by)?;
+        let (requested_by_principal, mut to_profile) =
+            ProfileStore::get(friend_request.requested_by)?;
         to_profile
             .relations
             .insert(friend_request.to, RelationType::Friend.to_string());
+
+        ProfileStore::update(caller(), caller_profile)?;
+        ProfileStore::update(requested_by_principal, to_profile)?;
 
         NotificationCalls::notification_accept_or_decline_friend_request(
             (friend_request_id, friend_request),
             true,
         )?;
+
         Ok(FriendRequestStore::remove(friend_request_id))
     }